player: Increase MIN_COMPLETED_PIECES to avoid failing to play some torrents
Some torrents (as eztv --> Suits --> S3E11) fail to play if we don't download
enough of the file before we play it. And is not that it stalls till it
downloads more, that wouldn't be too bad. What happens is that xbmc doesn't
think it is a video at all, so it fails and stops. And when it stops,
torrent2http stops too and basically you can't see the episode.

The problem is just that we didn't download enough of the file, as can be easily
reproduced using torrent2http manually. Also, the number 3 is chosen with this
rationale: 1 was not enough, 2 was not enough, 3 was enough. So is the smallest
that I tried that worked for at least for eztv --> Suits --> S3E11.

Maybe in future patches we can increase this number only for "affected"
torrents, or even try to play and if it fails continue waiting so not all
torrents need to wait for the same percentage to be complete.

Fixes #135