Wait untill 20MB (by default) are downloaded before playing
This heuristic seems to work better. Currently some videos eztv series (suits
S3E11 for example) fail to play if we don't download enough of the file before
we play it. And is not that it stalls till it downloads more, that wouldn't be
too bad. What happens is that xbmc doesn't think it is a video at all, so it
fails and stops. And when it stops, torrent2http stops too and basically you
can't see the episode.

The problem is just that we didn't download enough of the file, as can be easily
reproduced using torrent2http manually and trying to connect from xbmc, vlc,
mplayer or your favorite player.

Using the percentage of completed pieces is not too accurate to estimate if is
enough to play it. And, when increasing the percentage of pieces downloading to
make it work on most videos, the number ended up being too big for movies (i.e.
to download a lot) making the user wait for too long before the movie started.

Using a fixed amount of MB seems to work better in practice.

Fixes #135