Wait till stream is playable before playing
mplayer is run trying to connect to the stream. Once mplayer can connect to the
stream, we are good to go. We try with mplayer after downloading some MBs
because we want to download them anyways to make the playback smooth.

To simplify the code, the xbmc.sleep() is moved to the beggining of the loop and
whenever a condition is not reched, we just do "continue" and that makes the
sleep on the next iteration.

Also, this fixes some problems on some EZTV series that need more video to
download to play succesfully.